requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-anthropic>=0.0.1
langchain-huggingface>=0.0.1
langchain-chroma>=0.0.1
python-dotenv>=1.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
youtube-transcript-api>=0.6.1
chromadb>=0.4.22
sentence-transformers>=4.0.1
anthropic>=0.18.1
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-magic>=0.4.27 